pyproject.toml
[project]
name = "myshelf"
version = "0.1.0"
description = ""
authors = [
    {name = "Tamir Dingjan",email = "tamir.dingjan@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask (>=3.1.1,<4.0.0)",
    "pyaml (>=25.7.0,<26.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "requests (>=2.32.4,<3.0.0)"
]

[tool.poetry]
packages = [{include = "myshelf", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
